Move ChatInterface to chat_interface.py and clean imports
